fix update test
