+++ crates/dwn-core/src/message/record_id.rs
@@ -12,7 +12,7 @@ pub struct RecordIdGeneration {
 }
 
 impl Descriptor {
-    pub fn compute_record_id(&self) -> Result<String, CidGenerationError> {
+    pub fn compute_entry_id(&self) -> Result<String, CidGenerationError> {
         let generator = RecordIdGeneration {
             descriptor_cid: compute_cid_cbor(self)?,
         };
